+++ src/main/java/com/bankersCalculator/server/advice/loanAdvice/dto/api/LoanAdviceServiceRequest.java
@@ -7,19 +7,21 @@ import com.bankersCalculator.server.housingInfo.rentTransactionInquiry.common.Re
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class LoanAdviceServiceRequest {
 
 
     // 고객 정보
-    private Long rentalDeposit; // 임차보증금
-    private Long monthlyRent;   // 월세
-    private Long cashOnHand;    // 보유현금
+    private BigDecimal rentalDeposit; // 임차보증금
+    private BigDecimal monthlyRent;   // 월세
+    private BigDecimal cashOnHand;    // 보유현금
     private Integer age;    // 만나이
     private MaritalStatus maritalStatus; // 혼인상태
-    private Long annualIncome;  //연소득
-    private Long spouseAnnualIncome;    // 배우자연소득
+    private BigDecimal annualIncome;  //연소득
+    private BigDecimal spouseAnnualIncome;    // 배우자연소득
     private ChildStatus childStatus;    // 자녀상태
     private Boolean hasNewborn; // 신생아여부
     private HouseOwnershipType houseOwnershipType; // 주택소유형태
