+++ src/main/java/com/bankersCalculator/server/advice/loanAdvice/entity/LoanAdviceResult.java
@@ -5,6 +5,7 @@ import com.bankersCalculator.server.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -25,21 +26,21 @@ public class LoanAdviceResult {
     private String loanProductName;             // 대출 상품명
     private String loanProductCode;             // 대출 상품코드
 
-    private Long possibleLoanLimit;           // 가능한 대출 한도
-    private Double expectedLoanRate;            // 예상 대출 금리
+    private BigDecimal possibleLoanLimit;           // 가능한 대출 한도
+    private BigDecimal expectedLoanRate;            // 예상 대출 금리
 
-    private Long totalRentalDeposit;            // 총 임대 보증금
-    private Long loanAmount;                    // 대출 금액
-    private Long ownFunds;                      // 소요 자기 자금
+    private BigDecimal totalRentalDeposit;            // 총 임대 보증금
+    private BigDecimal loanAmount;                    // 대출 금액
+    private BigDecimal ownFunds;                      // 소요 자기 자금
 
-    private Long monthlyInterestCost;           // 월 이자 비용
-    private Long monthlyRent;                   // 월 임대료
+    private BigDecimal monthlyInterestCost;           // 월 이자 비용
+    private BigDecimal monthlyRent;                   // 월 임대료
 
-    private Long opportunityCostOwnFunds;       // 기회 비용
-    private Double depositInterestRate;         // 예금 이자율
+    private BigDecimal opportunityCostOwnFunds;       // 기회 비용
+    private BigDecimal depositInterestRate;         // 예금 이자율
 
-    private Long guaranteeInsuranceFee;         // 보증 보험료
-    private Long stampDuty;                     // 인지세
+    private BigDecimal guaranteeInsuranceFee;         // 보증 보험료
+    private BigDecimal stampDuty;                     // 인지세
 
     @Column(length = 4000)
     private String recommendationReason;        // 추천 이유
@@ -56,12 +57,12 @@ public class LoanAdviceResult {
     private String rentalLoanGuide;             // 임대 대출 가이드
 
     public static LoanAdviceResult create(User user, String loanProductName,
-                                          String loanProductCode, Long possibleLoanLimit,
-                                          Double expectedLoanRate, Long totalRentalDeposit,
-                                          Long loanAmount, Long ownFunds,
-                                          Long monthlyInterestCost, Long monthlyRent,
-                                          Long opportunityCostOwnFunds, Double depositInterestRate,
-                                          Long guaranteeInsuranceFee, Long stampDuty,
+                                          String loanProductCode, BigDecimal possibleLoanLimit,
+                                          BigDecimal expectedLoanRate, BigDecimal totalRentalDeposit,
+                                          BigDecimal loanAmount, BigDecimal ownFunds,
+                                          BigDecimal monthlyInterestCost, BigDecimal monthlyRent,
+                                          BigDecimal opportunityCostOwnFunds, BigDecimal depositInterestRate,
+                                          BigDecimal guaranteeInsuranceFee, BigDecimal stampDuty,
                                           String recommendationReason, List<RecommendedProduct> recommendedProducts,
                                           List<Bank> availableBanks, String rentalLoanGuide) {
         return LoanAdviceResult.builder()
